split CPU
